Remove check on SENTRY_ENV
Removes the check that the `SENTRY_ENV` matches or contains the
`RAILS_ENV` when deciding whether to initialize and configure Raven.

Also remove `SENTRY_ENV` from `config/templates/application.yml.erb`.

The presence or absence of `SENTRY_DSN_BACKEND` is sufficient to decide
whether to initialize Raven or not.

Sentry also has a `SENTRY_ENVIRONMENT` variable that can be used to set
the infrastructure environment independently of `RAILS_ENV` so
`SENTRY_ENV` is confusing.